pyproject.toml
[tool.poetry]
name = "ethanrosenthal"
version = "0.1.0"
description = ""
authors = ["EthanRosenthal <ethanrosenthal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyGithub = "^1.55"
pandas = "^1.4.3"
tabulate = "^0.8.10"

[tool.poetry.dev-dependencies]
ipykernel = "^6.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
